+++ LabBuilder/dsclibrary/MEMBER_DFSHUB.DSC.ps1
@@ -22,7 +22,7 @@ Configuration MEMBER_DFSHUB
     Import-DscResource -ModuleName ComputerManagementDsc
     Import-DscResource -ModuleName DFSDsc
     Import-DscResource -ModuleName StorageDsc
-    Import-DscResource -ModuleName xNetworking
+    Import-DscResource -ModuleName NetworkingDsc
 
     Node $AllNodes.NodeName {
         # Assemble the Local Admin Credentials