+++ LabBuilder/dsclibrary/MEMBER_JENKINS.DSC.ps1
@@ -17,7 +17,7 @@ Configuration MEMBER_JENKINS
     Import-DscResource -ModuleName 'PSDesiredStateConfiguration'
     Import-DscResource -ModuleName ComputerManagementDsc
     Import-DscResource -ModuleName cChoco
-    Import-DscResource -ModuleName xNetworking
+    Import-DscResource -ModuleName NetworkingDsc
 
     Node $AllNodes.NodeName {
         # Assemble the Local Admin Credentials