+++ LabBuilder/dsclibrary/modules/MyDSCResources/xCertAuthorityServer/xCertAuthorityServer.DSC.Schema.psm1
@@ -28,7 +28,7 @@ Configuration MEMBER_ROOTCA
     Import-DscResource -ModuleName ComputerManagementDsc
     Import-DscResource -ModuleName ActiveDirectoryCSDsc
     Import-DscResource -ModuleName xPSDesiredStateConfiguration
-    Import-DscResource -ModuleName xNetworking
+    Import-DscResource -ModuleName NetworkingDsc
     Node $AllNodes.NodeName {
         # Assemble the Local Admin Credentials
         if ($Node.LocalAdminPassword) {
@@ -225,21 +225,21 @@ Configuration MEMBER_ROOTCA
             }
 
             # Enable Online Responder FireWall rules so we can remote manage Online Responder
-            xFirewall OnlineResponderFirewall1
+            Firewall OnlineResponderFirewall1
             {
                 Name = "Microsoft-Windows-OnlineRevocationServices-OcspSvc-DCOM-In"
                 Enabled = "True"
                 DependsOn = "[ADCSOnlineResponder]ConfigOnlineResponder"
             }
 
-            xFirewall OnlineResponderirewall2
+            Firewall OnlineResponderirewall2
             {
                 Name = "Microsoft-Windows-CertificateServices-OcspSvc-RPC-TCP-In"
                 Enabled = "True"
                 DependsOn = "[ADCSOnlineResponder]ConfigOnlineResponder"
             }
 
-            xFirewall OnlineResponderFirewall3
+            Firewall OnlineResponderFirewall3
             {
                 Name = "Microsoft-Windows-OnlineRevocationServices-OcspSvc-TCP-Out"
                 Enabled = "True"