+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   Server 2016 domain with a enterprise root CA.
 - Correct certificate authority DSC Resources with ADCSCertificationAuthority
   to be IsSingleInstance.
+- Convert xNetworking to NetworkingDsc.
 
 ## 0.8.4.1160